+++ CHANGELOG.md
@@ -2,3 +2,10 @@
 
 ## v1.0.0 (31/3/2021)
 - First release
+
+<br>
+
+## v2.0.0 (date here)
+- Added the `list` command
+- Added the `config` command
+- Added support for gitignores and licenses